Final Class Push
I am pushing what I have. which sadly isn't a lot. I really ended up struggling with this class at the end. I broke my project 3 beyond repair when trying to make a dropdown menu, resulting in a huge loss of confidence to finish the rest of this class. I felt as If i had gone back to the beginning again and had to redo the entire project which i couldn't bear to do. Finally I created the project 4 wireframe and placed it in the image folder with my fonts also ready. Thats about all i got. After reading through the information I sort of had a panic attack, was afraid of wasting hours on something that would break and as a result getting the project started didn't really happen. I have to leave for california tomorrow morning and have at this point accepted I will be retaking this class next semester and hopefully can get on top of it. thanks for the semseter Justine and good luck with your health